Reduce verbose or redundant code constructs (#413)
- removes unnecessary code
- simplifies code
- adds missing Javadoc along the road

+++ dkpro-jwpl-revisionmachine/src/main/java/org/dkpro/jwpl/revisionmachine/common/util/MathUtilities.java
@@ -53,7 +53,7 @@ public class MathUtilities
      */
     public static double percentPlus(final double a, final double b)
     {
-        return round((double) a / (double) (a + b));
+        return round(a / (a + b));
     }
 
     /**
